pyproject.toml
[project]
name = "receipt-ai-expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "streamlit>=1.47.1",
    "uvicorn[standard]>=0.35.0",
]
